scripts/pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["haxscramper <haxscramper@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pypdf2 = "^3.0.1"
newspaper3k = "^0.2.8"
ebooklib = "^0.18"
lxml-html-clean = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
